+++ Public/JBOD/Get-SMARTInfo.ps1
@@ -1,6 +1,6 @@
 function Get-SMARTInfo {
     param(
-        $CDIPath = ""
+        $CDIPath = "C:\ProgramData\EasyRaidCheck\CrystalDiskInfo\DiskInfo64.exe"
     )
     
     $CDIExecutable = $CDIPath
@@ -13,16 +13,18 @@ function Get-SMARTInfo {
         Write-Error "An error occurred: $_"
     }
     
-    $warning = $false
-    # Define an array to store the drive objects
     $smartalldrives = @()
-    $drive = New-Object -TypeName PSObject
+
+    $drive = $null
 
     Get-Content -Path "C:\ProgramData\EasyRaidCheck\Crystaldiskinfo\DiskInfo.txt" | ForEach-Object {
-        # Check if the line contains drive information
         if ($_ -match 'Model : (.*)') {
-            # Add collected information to the drive object
-            $drive | Add-Member -MemberType NoteProperty -Name "Model" -Value $matches[1]
+            if ($drive) {
+                $smartalldrives += $drive
+            }
+            $drive = New-Object PSObject -Property @{
+                Model = $matches[1]
+            }
         }
         elseif ($_ -match 'Firmware : (.*)') {
             $drive | Add-Member -MemberType NoteProperty -Name "Firmware" -Value $matches[1]
@@ -30,32 +32,40 @@ function Get-SMARTInfo {
         elseif ($_ -match 'Serial Number : (.*)') {
             $drive | Add-Member -MemberType NoteProperty -Name "Serial Number" -Value $matches[1]
         }
+        elseif ($_ -match 'Disk Size : (.*)') {
+            $sizevalue = $matches[1] -replace '\s*\(.*\)', ''
+            $drive | Add-Member -MemberType NoteProperty -Name "Disk Size" -Value $sizevalue
+        }
         elseif ($_ -match 'Interface : (.*)') {
-            $drive | Add-Member -MemberType NoteProperty -Name "Interface" -Value $matches[1]
+            $interfacevalue = $matches[1] -replace '\s*\(.*\)', ''
+            $drive | Add-Member -MemberType NoteProperty -Name "Interface" -Value $interfacevalue
         }
         elseif ($_ -match 'Health Status : (.*)') {
             $drive | Add-Member -MemberType NoteProperty -Name "Health Status" -Value $matches[1]
+            if ($drive.'Health Status' -notmatch '\bGood\b' -and $null -ne $drive.'Health Status' -and $drive.'Health Status' -notmatch '\bUnknown\b') {
+                $drive | Add-Member -MemberType NoteProperty -Name "RowColour" -Value 'danger'
+            } else {
+                $drive | Add-Member -MemberType NoteProperty -Name "RowColour" -Value 'success'
+            }
         }
         elseif ($_ -match 'Power On Hours : (.*)') {
             $drive | Add-Member -MemberType NoteProperty -Name "Power On Hours" -Value $matches[1]
         }
         elseif ($_ -match 'Temperature : (.*)') {
-            $drive | Add-Member -MemberType NoteProperty -Name "Temperature" -Value $matches[1]
+            $temperaturevalue = $matches[1] -replace '\s*\(.*\)', ''
+            $drive | Add-Member -MemberType NoteProperty -Name "Temperature" -Value $temperaturevalue
         }
-        # Add other properties similarly
-        
         elseif ($_ -match 'Drive Letter : (.*)') {
-            # Add the drive letter to the existing drive object
-            $drive | Add-Member -MemberType NoteProperty -Name "DriveLetter" -Value $matches[1]
-
-            # Add the completed drive object to the array
+            $drive | Add-Member -MemberType NoteProperty -Name "Drive Letter" -Value $matches[1]
             $smartalldrives += $drive
-
-            # Reset the drive object for the next drive
-            $drive = New-Object -TypeName PSObject
+            $drive = $null
         }
     }
-    # Output the array of drive objects
-    $smartFailedDrives = $smartAllDrives | Where-Object { $_."Health Status" -notlike "*Good*" }
+
+    if ($drive) {
+        $smartalldrives += $drive
+    }
+
+    $smartFailedDrives = $smartalldrives | Where-Object { $_."Health Status" -notlike "*Good*" }
     return $smartalldrives, $smartFailedDrives
 }